Create Rankup toy game
Create a Rankup toy game that four players play one deck of cards, with no Jokers, no table cards. In this toy game, each action is just a single card, thus the action space is much smaller. Also the rules to decide legal actions are much easier to implement than that in a real Rankup game. Additonally, players don't call for lord suit, instead a lord card (suit and rank) is drawn randomly for each game in the training and testing runs.